cloudrun/agent_trading/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
google-cloud-storage>=2.14
httpx>=0.27
